Initial Commit - Base Anisble and Terraform Files
Initial files for setting up the 2GB linode using Terraform. Ansible will be used for deployment, but for now its just a copy of the files used for "Infinito27.com". Also, the hosts file in the ansible folder right now its a placeholder file. In the future the file will be ignored in order to use the placeholder always.