Nginx Proxy
Included a new container: Nginx Proxy. Really useful. A light NGINX container thats acting as an automated proxy for containers found in the docker compose stack. What I mean by this is that the proxy container will "proxy" or route requests to different containers depending on the value of their "VIRTUAL_HOST" environment variable. This means that, for example, if a container receives a request for "eddysanoli.com", it will redirect that traffic to the container that, in its docker compose declaration has a "VIRTUAL_HOST" variable equal to "eddysanoli.com". Very interesting container.