Bugfix: Removed try_files instruction
The proxy was trying to use a try files instruction, meaning that if no "index.html" file is found inside of "/var/nginx/www/" it will display a 404 error. Since I deleted the "index.html" files used by the NGINX container, the eddysanoli file was displaying a 404 error. Removed the instruction to allow the container to proxy the connection to the Next.js container.