Deployment Workflow
Decided to include a new github workflow that updates the website files each time a push to master is done. However, since app is running inside of a container, the "npm install" and "npm run build" commands would need to go inside the container. To tackle this problem I just included Node in the server itself that's running the containers. Since the container's files are mounted to the server's local filesystem, we can run the "npm" commands in the server and the changes will transfer to the inside of the container. To test this, I included a small change in the website's landing page and see if it works when pushing to master. I however, have not managed to install Node on the server. For whathever reason, my commands are now failing, so I will try to get a fix working as soon as possible.