Bugfix: Node Modules and .next not Mounted Volumes
To make the deployment workflows a bit easier to implement, I decided to install NVM in the server, use it to install node and then run the "npm run build" command on the server. I assumed that both the ".next" and "node_modules" folders were mounted as a volume to the server, meaning that the changes done in the server would be reflected in the internal server, however, that is not the case. I forgot that I skipped bothe the ".next" and "node_modules" folders from being a volume. So now Im re-mounting them to have my workflow work as expected.