Grafana + Loki
Forgot about the fact that each time I send a provisioning signal to the site, I have to renew my SSL certificates, causing them to renew AND you only get 5 renewals per week. So... I'm stuck. My site is not functional anymore so I added a couple of things on top meanwhile to test if they work: Grafana and Loki to have better access to internal logs of the application.

Changes:
- Included a new /grafana route for eddysanoli.com that should let me visit the grafana instance running in the docker compose stack
- Included a new "bot.eddysanoli.com" host for the nginx container
- Included containers for Loki, Promtail and Grafana in the cluster. I run grafana as root
- Ignored the ".next" files for the website
- Ignored the ".env" file used to store the environment variables used in ansible/docker-compose.yaml